+++ src/stdlib/atomic0.mli
@@ -1,0 +1,11 @@
+(*_********************************************************************************)
+(*_  pplumbing - Utility libraries to use with [pp]                               *)
+(*_  SPDX-FileCopyrightText: 2024-2026 Mathieu Barbin <mathieu.barbin@gmail.com>  *)
+(*_  SPDX-License-Identifier: MIT                                                 *)
+(*_********************************************************************************)
+
+include module type of struct
+  include Stdlib.Atomic
+end
+
+val set_temporarily : 'a t -> 'a -> f:(unit -> 'b) -> 'b