+++ src/stdlib/stdlib0.ml
@@ -4,6 +4,7 @@
 (*  SPDX-License-Identifier: MIT                                                 *)
 (*********************************************************************************)
 
+module Atomic = Atomic0
 module Code_error = Code_error0
 module Dyn = Dyn0
 module Err = Err0