+++ test/cram/command_test.ml
@@ -86,11 +86,10 @@ let print_styles_cmd =
        ]
      in
      let print_styled pp =
-       (match Err.color_mode () with
-        | `Never -> Format.printf "%a%!" Pp.to_fmt pp
-        | `Auto -> Pp_tty.print pp [@coverage off]
-        | `Always -> print_string (Pp_tty.to_string pp));
-       print_char '\n'
+       let fmt = Format.std_formatter in
+       if Err.should_enable_color Unix.stdout then Pp_tty.pp fmt pp else Pp.to_fmt fmt pp;
+       Format.pp_print_newline fmt ();
+       Format.pp_print_flush fmt ()
      in
      List.iter styles ~f:(fun (name, style) ->
        print_styled (Pp_tty.tag style (Pp.verbatim name)));
@@ -137,10 +136,25 @@ let emit_error_cmd =
      Err.error [ Pp.text "error message." ])
 ;;
 
+let log_styled_cmd =
+  Command.make
+    ~summary:"Emit a styled log message."
+    (let+ () = Log_cli.set_config () in
+     Log.app (fun () ->
+       [ Pp.concat
+           ~sep:Pp.space
+           [ Pp.text "Hello"
+           ; Pp_tty.tag (Ansi_styles [ `Fg_blue ]) (Pp.verbatim "Colored")
+           ; Pp.text "World!"
+           ]
+       ]))
+;;
+
 let main =
   Command.group
     ~summary:"Test pplumbing libs from the command line."
     [ "emit-error", emit_error_cmd
+    ; "log-styled", log_styled_cmd
     ; "logs", logs_cmd
     ; "print-styles", print_styles_cmd
     ; "sexp-and-dyn", sexp_and_dyn_cmd