+++ test/pplumbing-err/test__color_mode.ml
@@ -5,18 +5,16 @@
 (*********************************************************************************)
 
 let%expect_test "color_mode" =
-  List.iter Err.Color_mode.all ~f:(fun color_mode ->
-    Err.Private.color_mode := color_mode;
-    let color_mode' = Err.color_mode () in
-    require_equal (module Err.Color_mode) color_mode color_mode';
-    print_endline (Sexp.to_string_hum (Err.Color_mode.sexp_of_t color_mode')));
+  List.iter [ `Always; `Never ] ~f:(fun color_mode ->
+    Atomic.set_temporarily Err.Private.color_mode color_mode ~f:(fun () ->
+      let color_mode' = Err.color_mode () in
+      require_equal (module Err.Color_mode) color_mode color_mode';
+      print_endline (Sexp.to_string_hum (Err.Color_mode.sexp_of_t color_mode'))));
   [%expect
     {|
-    Auto
     Always
     Never
     |}];
-  Err.Private.color_mode := `Auto;
   ()
 ;;